add base ProductDetailsComponent
