+++ src/components/product/Product.component.tsx
@@ -1,37 +1,15 @@
-import React, { useContext } from 'react';
+import React from 'react';
 
-import { AppContext } from '@/context/AppContext.context.tsx';
+import { PageNotFoundComponent } from '@/components/page-not-found/PageNotFound.component.tsx';
+import { ProductDetailsComponent } from '@/components/product-details/ProductDetails.component.tsx';
 import { ProductDataProviderComponent } from '@/data-providers/ProductDataProvider.component.tsx';
-import type { AppContextProps } from '@/interfaces/AppContextProps.interface.tsx';
 
 interface ProductProps {
     id: number;
 }
 
-export const ProductComponent: React.FC<ProductProps> = ({ id }) => {
-    const appContext: AppContextProps = useContext(AppContext);
-
-    const returnToProducts = () => {
-        appContext.setRoutePath('#products');
-    };
-
-    return (
-        <ProductDataProviderComponent id={id}>
-            {({ product }) => (
-                <>
-                    <button onClick={returnToProducts}>Back</button>
-                    {product ? (
-                        <div className="product">
-                            <h2>{product.title}</h2>
-                            <img src={product.images[0]} alt={product.title} />
-                            <p>{product.description}</p>
-                            <p>Price: ${product.price}</p>
-                        </div>
-                    ) : (
-                        <p>Product not found</p>
-                    )}
-                </>
-            )}
-        </ProductDataProviderComponent>
-    );
-};
+export const ProductComponent: React.FC<ProductProps> = ({ id }) => (
+    <ProductDataProviderComponent id={id}>
+        {({ product }) => (product ? <ProductDetailsComponent {...product} /> : <PageNotFoundComponent />)}
+    </ProductDataProviderComponent>
+);
